% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R, R/lab3-package.R
\docType{package}
\name{euclidean}
\alias{euclidean}
\alias{euclidean-package}
\title{The euclidean algorithm that finds the
greates common divisor of two integers and Dijkstra's shortest path algorithm in a graph..}
\usage{
euclidean(firstNumber, secondNumber)
}
\arguments{
\item{firstNumber}{A number.}

\item{secondNumber}{A number.}
}
\value{
The value of euclidean algorithm calculation \code{firstNumber} and \code{secondNumber}.
}
\description{
This package provides two categories of important functions:
euclidean and dijkstra.
}
\section{euclidean functions finds the greates common divisor of two integers and Dijkstra's shortest path algorithm in a graph}{
NA
}

\section{dijkstra function takes a graph and an initial node and calculates the shortest path from the initial node to every other node in the graph.}{
NA
}

\examples{
euclidean(100,1000 )
}
\seealso{
\code{\link{https://en.wikipedia.org/wiki/Euclidean_algorithm}}
}
