\name{Lab03-package}
\alias{Lab03-package}
\alias{Lab03}
\docType{package}
\title{
\packageTitle{Lab03}
}
\description{
\packageDescription{Lab03}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab03}
\packageIndices{Lab03}
This package include two basic function. Euclidean which generates the greatest common divisor of two integers and Dijkstra which finds the shortest path in the graph
}
\author{
\packageAuthor{Lab03}

Maintainer: \packageMaintainer{Lab03}
}

\keyword{ package }

\examples{
euclidean(10,100) = 10
}
