% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab03-package.R
\docType{package}
\name{lab03}
\alias{lab03}
\title{lab03: A package intended for the R package learning and Statistic computation}
\description{
The Lab03 package provides 2 important functions:
Euclidean Distance and Dijkstra's Algorithm.
}
\section{lab03 functions}{

The lab functions are named dijkstra and euclidean
}

