% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_values.R
\docType{data}
\name{random_values}
\alias{random_values}
\title{Random 20 Numerical Values.}
\format{A data frame with 20 random variables:
\describe{
  \item{Value_One}{First Numerical Value}
  \item{Value_Two}{Second Numerical Value}

}}
\source{
Generated Random Values for  test Gratest common divisor. \url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\usage{
random_values
}
\description{
A dataset containing 20 numberical values for test Euclidean Algorithm.
}
\keyword{datasets}
