% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_graphs.R
\docType{data}
\name{sample_graphs}
\alias{sample_graphs}
\title{Random 10 vector values lenght of 18}
\format{A data frame with 18 rows and 10 columns:
\describe{
  \item{V1}{A vector inside the wiki_graph dataframe containing the first edge.}
  \item{V2}{A vector inside the wiki_graph dataframe containing the second edge.}
  \item{V3}{A vector inside the wiki_graph dataframe containing the second edge.}
  \item{V4}{A vector inside the wiki_graph dataframe containing the second edge.}
  \item{V5}{A vector inside the wiki_graph dataframe containing the second edge.}
  \item{V6}{A vector inside the wiki_graph dataframe containing the second edge.}
  \item{V7}{A vector inside the wiki_graph dataframe containing the second edge.}
  \item{V8}{A vector inside the wiki_graph dataframe containing the second edge.}
  \item{V9}{A vector inside the wiki_graph dataframe containing the second edge.}
  \item{V10}{A vector inside the wiki_graph dataframe containing the second edge.}

  ...
}}
\source{
\url{https://en.wikipedia.org/wiki/Graph/}
}
\usage{
sample_graphs
}
\description{
Random 10 vector values lenght of 18
}
\keyword{datasets}
