% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{A dataset called wiki_graph used to check the output of dijkstra algorithm.}
\format{A data frame with 54 rows and 3 variables:
\describe{
  \item{v1}{A vector inside the wiki_graph dataframe containing the first edge.}
  \item{v2}{A vector inside the wiki_graph dataframe containing the second edge.}
  \item{w}{A vector inside the wiki_graph dataframe containing the weight between different edges.}
  ...
}}
\source{
\url{https://en.wikipedia.org/wiki/Graph/}
}
\usage{
wiki_graph
}
\description{
A dataset called wiki_graph used to check the output of dijkstra algorithm.
}
\keyword{datasets}
