% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijikstra}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{a graph as data frame.}

\item{init_node}{the initial node.}
}
\value{
The shortest path of \code{graph} with the initial node \code{init_node}
}
\description{
The \code{Dijkstra} algorithm calculates the shortest path of a graph with an initial node.
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
\references{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
