% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra_function.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's Algorithm}
\usage{
dijkstra(wiki_graph, init_node)
}
\arguments{
\item{wiki_graph}{A data frame representing the graph, with three columns:
\describe{
  \item{v1}{The starting node of an edge.}
  \item{v2}{The ending node of an edge.}
  \item{weight}{A numeric value representing the weight of the edge.}
}}

\item{init_node}{The starting node (must match a node in \code{graph}).}
}
\value{
A numeric vector where each element gives the shortest distance
from \code{init_node} to that node. Nodes that are unreachable will have value \code{Inf}.
}
\description{
This function implements **Dijkstra's algorithm** to compute the path
distances from a starting node to all other nodes in a weighted graph.
}
\details{
Dijkstra's algorithm is a greedy algorithm that repeatedly selects the unvisited
node with the smallest known distance, updates the distances of its neighbors,
and marks it as visited. The algorithm continues until all nodes are visited or
the smallest known distance among unvisited nodes is infinite.
}
\examples{
# Example graph
wiki_graph <- data.frame(
  v1     = c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
  v2     = c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
  weight = c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)
)

dijkstra(wiki_graph, init_node = 1) #returns [1] 0 7 9 20 20 11
dijkstra(wiki_graph, init_node = 3) #returns [1] 9 10 0 11 11 2

}
\references{
Wikipedia: Dijkstra's Algorithm - https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
