% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{dijkstra}
\usage{
dijkstra(graph,init_node)
}
\arguments{
\item{graph}{data frame}

\item{init_node}{number}
}
\value{
dijkstra(graph,init_node)
}
\description{
recieve graph and initial node and calculate shortest path.
}
\details{
dijkstra which calculates the shortest path from initial node in a graph.
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\seealso{
\url{https://en.wikipedia.org/wiki/Graph_(mathematics}
}
\author{
Hamed
}
