% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra algorithm}
\source{
\href{https://en.wikipedia.org/wiki/Dijkstra_algorithm}{Dijkstra algorithm}
}
\usage{
dijkstra(wiki_graph, init_node)
}
\arguments{
\item{wiki_graph}{data.frame}

\item{init_node}{integer}
}
\value{
vector in which nth element is equal to the shortest path from initial node to nth vertex
}
\description{
Finding the shortest paths in the graph from initial node
}
\author{
Reuel, Martin, Vinay
}
