% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{The euclidean algorithm.}
\usage{
euclidean(firstNumber, secondNumber)
}
\arguments{
\item{firstNumber}{A number.}

\item{secondNumber}{A number.}
}
\value{
The value of euclidean algorithm calculation \code{firstNumber} and \code{secondNumber}.
}
\description{
The euclidean algorithm.
}
\examples{
euclidean(100,1000 )
}
