% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R, R/lab3-package.R
\docType{package}
\name{euclidean}
\alias{euclidean}
\alias{euclidean-package}
\title{The Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A Numerical value}

\item{b}{A Numerical Value}
}
\value{
Gratest common divior of given values
}
\description{
Euclidean Algorithm is an efficient method for computing the greatest common divisor (GCD) of two numbers, the largest number that divides both of them without leaving a remainder.

This package provides two categories of important functions:
euclidean and dijkstra.
}
\section{euclidean functions finds the greates common divisor of two integers and Dijkstra's shortest path algorithm in a graph}{
NA
}

\section{dijkstra function takes a graph and an initial node and calculates the shortest path from the initial node to every other node in the graph.}{
NA
}

\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\seealso{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
