% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm implementation}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{integer}

\item{b}{integer}
}
\value{
a
}
\description{
Euclidean Algorithm implementation
}
\examples{
euclidean(100, 1000)
}
