% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{euclidean}
\usage{
euclidean(b, a)
}
\arguments{
\item{b}{a number}

\item{a}{a number}
}
\value{
The GCD of \code{b} and \code{a}.
}
\description{
\code{euclidean} use the euclidean algorithm to calculate the Greatest Common Divisor (GCD) of 2 numbers.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
