% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labo3.R
\name{euclidean}
\alias{euclidean}
\title{It is used to find the greatest common devisier}
\usage{
euclidean(H1, H2)
}
\arguments{
\item{H1}{For first value}

\item{H2}{For second value}
}
\value{
it return the greatest common devisier
}
\description{
It is used to find the greatest common devisier
}
\examples{
euclidean(123612,13892347912)
euclidean(100,1000)
}
