% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{\itemize{
\item number (integer)
}}

\item{b}{\itemize{
\item number (integer)
}}
}
\value{
greatest common divisor
}
\description{
The euclidean algorithm finds the greatest common divisor of two integers.
See \url{https://en.wikipedia.org/wiki/Euclidean_algorithm} for more information.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
