% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{\itemize{
\item number (integer)
}}

\item{b}{\itemize{
\item number (integer)
}}
}
\value{
greatest common divisor
}
\description{
Euclidean
}
\examples{
euclidean(123612, 13892347912)
}
