% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean_function.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric integer scalar.}

\item{b}{A numeric integer scalar.}
}
\value{
An integer value representing the greatest common divisor
of \code{a} and \code{b}.
}
\description{
This function implements the **Euclidean algorithm** to compute the
greatest common divisor of two integers. The Euclidean algorithm
is based on the principle that the greatest common divisor of two numbers
also divides their remainder when one is divided by the other.
}
\details{
The function first checks that both inputs are integers.
Negative numbers are converted to their absolute values.
The algorithm repeatedly replaces the larger number with the remainder
of dividing the larger by the smaller until the remainder is zero.
At that point, the smaller number is the greatest common divisor.
}
\examples{
euclidean(123612, 13892347912)  # returns 4
euclidean(100, 1000) # returns 100

}
\references{
Euclidean Algorithm explanation: \url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
