% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean function, finds the greatest common divisor of two numbers}
\usage{
euclidean(value1, value2)
}
\arguments{
\item{value1}{An integer number must be introduced}

\item{value2}{An integer number must be introduced}
}
\value{
It will return the greatest common divisor between value 1 and value 2. An integer will be returned.
}
\description{
Euclidean function, finds the greatest common divisor of two numbers. Find Wikipedia docs \href{https://en.wikipedia.org/wiki/Euclidean}{Here}
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)

}
