% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{\title{Euclidean} 
Finds the greatest common divisor of two numbers
See the Wikipedia documentation \href{https://en.wikipedia.org/wiki/Euclidean}{Here}}
\usage{
euclidean(value1, value2)
}
\arguments{
\item{value1}{A number.}

\item{value2}{A number.}
}
\value{
Greatest common divisor.
}
\description{
\title{Euclidean} 
Finds the greatest common divisor of two numbers
See the Wikipedia documentation \href{https://en.wikipedia.org/wiki/Euclidean}{Here}
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
