% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Finds the greatest common divisor of two numbers}
\usage{
euclidean(value1, value2)
}
\arguments{
\item{value1}{A number.}

\item{value2}{A number.}
}
\value{
Greatest common divisor.
}
\description{
Finds the greatest common divisor of two numbers
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
