% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Finds the greatest common divisor of two numbers}
\usage{
euclidean(value1, value2)
}
\arguments{
\item{value1}{An integer number.}

\item{value2}{An integer number.}
}
\value{
Greatest common divisor - An integer
}
\description{
\code{euclidean}
}
\details{
Find Wikipedia docs \href{https://en.wikipedia.org/wiki/Euclidean}{Here}
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
