% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean shit}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{number}

\item{b}{number}
}
\value{
return GCD
}
\description{
Euclidean shit
}
\examples{
euclidean(10,100)
}
