% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Greatest Common Divisor of two numbers.}
\usage{
euclidean(num1, num2)
}
\arguments{
\item{num1}{A number.}

\item{num2}{A number.}
}
\value{
The Greatest Common Divisor
}
\description{
Find the Greatest Common Divisor of two given numbers
}
\details{
Should assert that the arguments are numeric scalars or integers.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\author{
Hamed
}
