% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{euclidean
euclidiean calculates the Greatest Common Divisor (GCD) for 2 numbers}
\source{
see more at: \url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\usage{
euclidean(b, a)
}
\arguments{
\item{b}{a number}

\item{a}{a number}
}
\value{
The GCD of \code{b} and \code{a}.
}
\description{
euclidean use the euclidean algorithm to calculcate the GCD.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
