% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean fcn.R
\name{euclidean}
\alias{euclidean}
\title{Finding the greatest common divisor (GCD)}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{integer}

\item{b}{integer}
}
\value{
an integer equal to GCD
}
\description{
This function finds the GCD by using modulus function and while loop
}
\author{
Reuel, Martin, Vinay
}
