% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{euclidean algorithm}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
greatest common divisor (GCD) of x and y
}
\description{
This function is a loop algorithm to find the greatest common
divisor (GCD) of two nonzero number x and y using euclidean algorithm. If the
lager number devided by the smaller number has a nonzero remainder,replace the
lager number with remainder.Reverse the step until the remainder is zero, then
the denominator is GCD of the original two numbers.
}
\examples{
euclidean(100,1000)
euclidean(1234,8888)
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
