\name{Lab03-package}
\alias{Lab03-package}
\alias{Lab03}
\docType{package}
\title{
\packageTitle{Lab03}
}
\description{
\packageDescription{Lab03}
}
\details{

The DESCRIPTION file:
\packageDescribtion{Lab03}


}
\author{
\packageAuthor{Lab03}

Maintainer: \packageMaintainer{Lab03}
}


\keyword{ package }

