\name{Lab03-package}
\alias{Lab03-package}
\alias{Lab03}
\docType{package}
\title{
\packageTitle{Lab03}
}
\description{
\packageDescription{Lab03}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab03}
\packageIndices{Lab03}
 An overview of how to use the package, including the most important functions 
}
\author{
\packageAuthor{Lab03}

Maintainer: \packageMaintainer{Lab03}
}
\references{
 Literature or other references for background information 
}
 Optionally other standard keywords, one per line, from file KEYWORDS in the R 
 documentation directory 
\keyword{ package }
\seealso{
Optional links to other man pages, e.g.
\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
simple examples of the most important functions
}
