% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Grap: edges and weights}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{v1}{node, egde pair 1}
\item{v2}{node, edge pair 2}
\item{w}{weight of the edge}
...
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
A dataset containing the edges and weights of a graph.
}
\keyword{datasets}
