% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph data frame}
\format{
## `wiki_graph`
A data frame with with three columns and as many rows a needed:
\describe{
  \item{v1}{The starting node of an edge.}
  \item{v2}{The ending node of an edge.}
  \item{weight}{A numeric value representing the weight of the edge.}
}
}
\source{
<https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm>
}
\usage{
wiki_graph
}
\description{
a sample data frame used to test the Dijkstra's Algorithm funtion
}
\keyword{datasets}
