% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{18 edges(v1 -v2 ) and weights .}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{node}
  \item{v2}{destination node}
  \item{w}{weight}
  ...
}
}
\source{

}
\usage{
wiki_graph
}
\description{
A dataset containing 18 edges and weights.
}
\keyword{datasets}
