% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki_graph}
\format{An object of class \code{data.frame} with 18 rows and 3 columns.}
\source{
\href{https://en.wikipedia.org/wiki/Graph}{wiki_graph}
}
\usage{
data(wiki_graph)
}
\description{
Creates a graph with 6 vertices and 18 edges
}
\keyword{datasets}
